langchain/requirements.txt
# LangChain/LangGraph Agentic Patterns Implementation
# Core dependencies for implementing all 170 patterns

# Core LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
langchain-openai>=0.0.5

# Additional LLM providers (optional)
langchain-anthropic>=0.0.1
langchain-google-genai>=0.0.5

# Memory and storage
langchain-chroma>=0.0.1
faiss-cpu>=1.7.4

# Tools and utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Testing and validation
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Monitoring and observability
langsmith>=0.0.70

# Additional utilities
pydantic>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
